+++ CHANGELOG.md
@@ -6,7 +6,11 @@ We use [semantic versioning](http://semver.org/):
 
 # Next Release
 
+<<<<<<< HEAD
+- [breaking change] certificates are now validated by default. Use `--insecure` to disable validation
+=======
 - [breaking change] remove `--detect-commit` option and make that the default behaviour
+>>>>>>> origin/master
 - [feature] allow users to provide self-signed certificates that should be trusted in a Java keystore
 - [feature] allow users to append lines to the default message